requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# Audio processing and transcription
openai-whisper==20231117
pyannote.audio==3.1.1
pydub==0.25.1
torch==2.1.1
torchaudio==2.1.1

# NLP for title suggestions
transformers==4.36.0
torch==2.1.1
scikit-learn==1.3.2
nltk==3.8.1

# File handling
Pillow==10.1.0

# API documentation
drf-yasg==1.21.7

# Environment variables
python-decouple==3.8
